Uniform navigator interface
Previously, we were overriding and deleting properties from the Navigator API in
order to modify / disable them. If a new property was added, or was non
consistent between devices / browser versions, those would be exposed
automatically. This commit creates a fake Navigator API and only exposes the
functionality we have wrapped and intentionally decided to expose.